+++ Assembly-CSharp/Engine/Source/Services/PostmanTeleportUtility.cs
@@ -1,6 +1,5 @@
 ﻿using Engine.Source.Commons;
 using Engine.Source.Settings;
-using UnityEngine;
 
 namespace Engine.Source.Services
 {
@@ -12,7 +11,7 @@ namespace Engine.Source.Services
       if ((double) to.magnitude > 40.0)
         return false;
       to.y = 0.0f;
-      return (double) Vector3.Angle(playerGameObject.transform.forward, to) < (double) (InstanceByRequest<GraphicsGameSettings>.Instance.FieldOfView.Value * GameCamera.Instance.Camera.aspect) / 2.0;
+      return (double) Vector3.Angle(playerGameObject.transform.forward, to) < InstanceByRequest<GraphicsGameSettings>.Instance.FieldOfView.Value * GameCamera.Instance.Camera.aspect / 2.0;
     }
   }
 }
