+++ Assembly-CSharp/SRFileUtil.cs
@@ -19,40 +19,40 @@ public static class SRFileUtil
   public static string GetBytesReadable(long i)
   {
     string str1 = i < 0L ? "-" : "";
-    double num1 = i < 0L ? (double) -i : (double) i;
+    double num1 = i < 0L ? -i : (double) i;
     string str2;
     double num2;
     if (i >= 1152921504606846976L)
     {
       str2 = "EB";
-      num2 = (double) (i >> 50);
+      num2 = i >> 50;
     }
     else if (i >= 1125899906842624L)
     {
       str2 = "PB";
-      num2 = (double) (i >> 40);
+      num2 = i >> 40;
     }
     else if (i >= 1099511627776L)
     {
       str2 = "TB";
-      num2 = (double) (i >> 30);
+      num2 = i >> 30;
     }
     else if (i >= 1073741824L)
     {
       str2 = "GB";
-      num2 = (double) (i >> 20);
+      num2 = i >> 20;
     }
     else if (i >= 1048576L)
     {
       str2 = "MB";
-      num2 = (double) (i >> 10);
+      num2 = i >> 10;
     }
     else
     {
       if (i < 1024L)
         return i.ToString(str1 + "0 B");
       str2 = "KB";
-      num2 = (double) i;
+      num2 = i;
     }
     double num3 = num2 / 1024.0;
     return str1 + num3.ToString("0.### ") + str2;
